+++ src/components/OnPremise/UnlockDeckDoorBottomSheet.tsx
@@ -67,7 +67,7 @@ const UnlockDeckDoorBottomSheet = ({
 
   return (
     <AppBottomSheet
-      contentContainerStyle={tw`flex flex-col items-center gap-4 px-6 pt-6 pb-4`}
+      contentContainerStyle={tw`flex flex-col items-center gap-4 px-6 pt-6`}
       style={style}
       onClose={onClose}
       {...(Platform.OS === 'android' && { animationConfigs: { duration: 300 } })}>